Remove TabRepo (part 1)
