+++ tabarena/tabarena/contexts/__init__.py
@@ -1,2 +1,1 @@
-from ._context_fetcher import get_context, get_subcontext, list_contexts, list_subcontexts
 from .context import BenchmarkContext
